[refactor] #10 TimeSlot 엔티티 STI 구조로 통합 및 파일 분리
